requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
requests
beautifulsoup4
python-whois
dnspython
tldextract
google-generativeai
python-dotenv
